+++ Sources/SwiftSDK/Persistence/DataStoreFactory.swift
@@ -24,7 +24,6 @@
     typealias CustomType = Any
     
     public var rt: EventHandlerForClass!
-    public private(set) var isOfflineAutoSyncEnabled = false
     
     private var entityClass: AnyClass
     private var tableName: String
@@ -67,7 +66,7 @@
         let entityDictionary = PersistenceHelper.shared.entityToDictionary(entity: entity)
         let wrappedBlock: ([String: Any]) -> () = { responseDictionary in
             let className = PersistenceHelper.shared.getClassNameWithoutModule(self.entityClass)
-            if let resultEntity = PersistenceHelper.shared.dictionaryToEntity(dictionary: responseDictionary, className: className) {
+            if let resultEntity = PersistenceHelper.shared.dictionaryToEntity(responseDictionary, className: className) {
                 responseHandler(resultEntity)
             }
         }
@@ -86,7 +85,7 @@
         let entityDictionary = PersistenceHelper.shared.entityToDictionary(entity: entity)
         let wrappedBlock: ([String: Any]) -> () = { responseDictionary in
             let className = PersistenceHelper.shared.getClassNameWithoutModule(self.entityClass)
-            if let resultEntity = PersistenceHelper.shared.dictionaryToEntity(dictionary: responseDictionary, className: className) {
+            if let resultEntity = PersistenceHelper.shared.dictionaryToEntity(responseDictionary, className: className) {
                 responseHandler(resultEntity)
             }
         }
@@ -124,7 +123,7 @@
             var resultArray = [Any]()
             for responseObject in responseArray {
                 let className = PersistenceHelper.shared.getClassNameWithoutModule(self.entityClass)
-                if let resultObject = PersistenceHelper.shared.dictionaryToEntity(dictionary: responseObject, className: className) {
+                if let resultObject = PersistenceHelper.shared.dictionaryToEntity(responseObject, className: className) {
                     resultArray.append(resultObject)
                 }
             }
@@ -138,7 +137,7 @@
             var resultArray = [Any]()
             for responseObject in responseArray {
                 let className = PersistenceHelper.shared.getClassNameWithoutModule(self.entityClass)
-                if let resultObject = PersistenceHelper.shared.dictionaryToEntity(dictionary: responseObject, className: className) {
+                if let resultObject = PersistenceHelper.shared.dictionaryToEntity(responseObject, className: className) {
                     resultArray.append(resultObject)
                 }
             }
@@ -148,19 +147,13 @@
     }
     
     public func findFirst(responseHandler: ((Any) -> Void)!, errorHandler: ((Fault) -> Void)!) {
-        let wrappedBlock: ([String: Any]) -> () = { responseDictionary in
-            let className = PersistenceHelper.shared.getClassNameWithoutModule(self.entityClass)
-            if let resultEntity = PersistenceHelper.shared.dictionaryToEntity(dictionary: responseDictionary, className: className) {
-                responseHandler(resultEntity)
-            }
-        }
-        persistenceServiceUtils.findFirstOrLastOrById(first: true, last: false, objectId: nil, queryBuilder: nil, responseHandler: wrappedBlock, errorHandler: errorHandler)
+        findFirst(queryBuilder: DataQueryBuilder(), responseHandler: responseHandler, errorHandler: errorHandler)
     }
     
     public func findFirst(queryBuilder: DataQueryBuilder, responseHandler: ((Any) -> Void)!, errorHandler: ((Fault) -> Void)!) {
         let wrappedBlock: ([String: Any]) -> () = { responseDictionary in
             let className = PersistenceHelper.shared.getClassNameWithoutModule(self.entityClass)
-            if let resultEntity = PersistenceHelper.shared.dictionaryToEntity(dictionary: responseDictionary, className: className) {
+            if let resultEntity = PersistenceHelper.shared.dictionaryToEntity(responseDictionary, className: className) {
                 responseHandler(resultEntity)
             }
         }
@@ -168,19 +161,13 @@
     }
     
     public func findLast(responseHandler: ((Any) -> Void)!, errorHandler: ((Fault) -> Void)!) {
-        let wrappedBlock: ([String: Any]) -> () = { responseDictionary in
-            let className = PersistenceHelper.shared.getClassNameWithoutModule(self.entityClass)
-            if let resultEntity = PersistenceHelper.shared.dictionaryToEntity(dictionary: responseDictionary, className: className) {
-                responseHandler(resultEntity)
-            }
-        }
-        persistenceServiceUtils.findFirstOrLastOrById(first: false, last: true, objectId: nil, queryBuilder: nil, responseHandler: wrappedBlock, errorHandler: errorHandler)
+        findLast(queryBuilder: DataQueryBuilder(), responseHandler: responseHandler, errorHandler: errorHandler)
     }
     
     public func findLast(queryBuilder: DataQueryBuilder, responseHandler: ((Any) -> Void)!, errorHandler: ((Fault) -> Void)!) {
         let wrappedBlock: ([String: Any]) -> () = { responseDictionary in
             let className = PersistenceHelper.shared.getClassNameWithoutModule(self.entityClass)
-            if let resultEntity = PersistenceHelper.shared.dictionaryToEntity(dictionary: responseDictionary, className: className) {
+            if let resultEntity = PersistenceHelper.shared.dictionaryToEntity(responseDictionary, className: className) {
                 responseHandler(resultEntity)
             }
         }
@@ -188,19 +175,13 @@
     }
     
     public func findById(objectId: String, responseHandler: ((Any) -> Void)!, errorHandler: ((Fault) -> Void)!) {
-        let wrappedBlock: ([String: Any]) -> () = { responseDictionary in
-            let className = PersistenceHelper.shared.getClassNameWithoutModule(self.entityClass)
-            if let resultEntity = PersistenceHelper.shared.dictionaryToEntity(dictionary: responseDictionary, className: className) {
-                responseHandler(resultEntity)
-            }
-        }
-        persistenceServiceUtils.findFirstOrLastOrById(first: false, last: false, objectId: objectId, queryBuilder: nil, responseHandler: wrappedBlock, errorHandler: errorHandler)
+        findById(objectId: objectId, queryBuilder: DataQueryBuilder(), responseHandler: responseHandler, errorHandler: errorHandler)
     }
     
     public func findById(objectId: String, queryBuilder: DataQueryBuilder, responseHandler: ((Any) -> Void)!, errorHandler: ((Fault) -> Void)!) {
         let wrappedBlock: ([String: Any]) -> () = { responseDictionary in
             let className = PersistenceHelper.shared.getClassNameWithoutModule(self.entityClass)
-            if let resultEntity = PersistenceHelper.shared.dictionaryToEntity(dictionary: responseDictionary, className: className) {
+            if let resultEntity = PersistenceHelper.shared.dictionaryToEntity(responseDictionary, className: className) {
                 responseHandler(resultEntity)
             }
         }
@@ -232,27 +213,18 @@
     }
     
     public func loadRelations(objectId: String, queryBuilder: LoadRelationsQueryBuilder, responseHandler: (([Any]) -> Void)!, errorHandler: ((Fault) -> Void)!) {
-        let wrappedBlock: ([[String: Any]]) -> () = { responseArray in
+        let wrappedBlock: ([Any]) -> () = { responseArray in
             var resultArray = [Any]()
             for responseObject in responseArray {
-                if responseObject["___class"] as? String == "Users",
-                    let userObject = ProcessResponse.shared.adaptToBackendlessUser(responseResult: responseObject) {
-                    resultArray.append(userObject as! BackendlessUser)
-                }
-                else if responseObject["___class"] as? String == "GeoPoint",
-                    let geoPointObject = ProcessResponse.shared.adaptToGeoPoint(geoDictionary: responseObject) {
-                    resultArray.append(geoPointObject)
-                }
-                else if responseObject["___class"] as? String == "DeviceRegistration" {
-                    let deviceRegistrationObject = ProcessResponse.shared.adaptToDeviceRegistration(responseResult: responseObject)
-                    resultArray.append(deviceRegistrationObject)
-                }
-                // addGeo
-                else if let relationType = queryBuilder.getRelationType() {
-                    let tableName = PersistenceHelper.shared.getClassNameWithoutModule(relationType)
-                    if let resultObject = PersistenceHelper.shared.dictionaryToEntity(dictionary: responseObject, className: tableName) {
+                if let dictResponse = responseObject as? [String : Any],
+                    let relationType = queryBuilder.getRelationType() {
+                    let className = PersistenceHelper.shared.getClassNameWithoutModule(relationType)
+                    if let resultObject = PersistenceHelper.shared.dictionaryToEntity(dictResponse, className: className) {
                         resultArray.append(resultObject)
                     }
+                }
+                else {
+                    resultArray.append(responseObject)
                 }
             }
             responseHandler(resultArray)