+++ Sources/SwiftSDK/Persistence/IDataStore.swift
@@ -51,7 +51,7 @@ protocol IDataStore {
     func addRelation(columnName: String, parentObjectId: String, whereClause: String?, responseHandler: ((Int) -> Void)!, errorHandler: ((Fault) -> Void)!)
     func deleteRelation(columnName: String, parentObjectId: String, childrenObjectIds: [String], responseHandler: ((Int) -> Void)!, errorHandler: ((Fault) -> Void)!)
     func deleteRelation(columnName: String, parentObjectId: String, whereClause: String?, responseHandler: ((Int) -> Void)!, errorHandler: ((Fault) -> Void)!)
-    func loadRelations(objectId: String, queryBuilder: LoadRelationsQueryBuilder, responseHandler: (([CustomType]) -> Void)!, errorHandler: ((Fault) -> Void)!)
+    func loadRelations(objectId: String, queryBuilder: LoadRelationsQueryBuilder, responseHandler: (([Any]) -> Void)!, errorHandler: ((Fault) -> Void)!)
     
     func initLocalDatabase(whereClause: String, responseHandler: (() -> Void)!, errorHandler: ((Fault) -> Void)!)
     func clearLocalDatabase()
@@ -63,5 +63,4 @@ protocol IDataStore {
     func saveEventually(entity: CustomType, callback: OfflineAwareCallback)
     func removeEventually(entity: CustomType)
     func removeEventually(entity: CustomType, callback: OfflineAwareCallback)
-    
 }